Refactor to also handle deferred intents
